+++ Assets/_Scripts/Units/UI/Menus/IMenuScreen.cs
@@ -1,0 +1,10 @@
+namespace Metroidvania.UI.Menus
+{
+    public interface IMenuScreen
+    {
+        bool menuEnabled { get; }
+
+        void ActiveMenu();
+        void DesactiveMenu();
+    }
+}